+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- new `auth` helper to exchange device tokens for session tokens
+- `session` synchronous factory for constructing clients from cached session tokens
+
 ## [8.4.0] - 2025-08-17
 
 ### Changed